+++ src/Doka.EntityFrameworkCore.SafeMigrations.MySql/Imports.cs
@@ -5,6 +5,7 @@ global using System.Diagnostics;
 global using System.Globalization;
 global using System.Linq;
 global using System.Text;
+global using System.Text.RegularExpressions;
 global using System.Threading;
 global using System.Threading.Tasks;
 global using Doka.EntityFrameworkCore.MySql;